+++ src/main/java/com/ath/adminefectivo/repositories/IDetallesLiquidacionCostoRepository.java
@@ -7,8 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
 import com.ath.adminefectivo.entities.DetallesLiquidacionCosto;
-import com.ath.adminefectivo.entities.ValoresLiquidados;
-import com.ath.adminefectivo.entities.ValoresLiquidadosFlatEntity;
 
 public interface IDetallesLiquidacionCostoRepository extends JpaRepository<DetallesLiquidacionCosto, Long>, 
 QuerydslPredicateExecutor<DetallesLiquidacionCosto>{
@@ -20,6 +18,6 @@ QuerydslPredicateExecutor<DetallesLiquidacionCosto>{
 	 * @author hector.mercado
 	 */
 	@Query(value = "SELECT * FROM detalles_liquidacion_costo WHERE ID_LIQUIDACION = ?1" , nativeQuery = true)
-	List<DetallesLiquidacionCosto> consultarPorIdLiquidacion(Long idLiquidacion);
+	List<DetallesLiquidacionCosto> consultarPorIdLiquidacion(Integer idLiquidacion);
 
 }
